refactor: 모듈 경로 변경
